--- src/components/OpsSidenav/style.scss
@@ -1,167 +1,0 @@
-@import "@shared/styles/_variables.scss";
-
-.sidebar {
-  padding-top: 12px;
-
-  ul.navigation {
-    list-style: none;
-    padding: 0;
-
-    li {
-      padding: 0 7px 0 49px;
-    }
-
-    .btn-link-nav {
-      text-decoration: none;
-      &:focus {
-        outline: -webkit-focus-ring-color auto 1px;
-      }
-    }
-
-    .item {
-      li {
-        a {
-          text-align: left;
-          color: #2b2b2b;
-          font-size: 14px;
-
-          &.btn-link-link {
-            color: $blue-lighten-5;
-            font-weight: 600;
-          }
-        }
-      }
-    }
-
-    li.separated {
-      margin-top: 50px;
-    }
-
-    li.subheader {
-      margin-top: 25px;
-      margin-bottom: 5px;
-      span {
-        font-size: 90%;
-        user-select: none;
-        color: $gray-lighten-1;
-      }
-    }
-
-    span.code-field {
-      margin-left: 5px;
-      font-size: 80%;
-      text-transform: uppercase;
-    }
-
-    li.submenu, li.previous-link {
-      display: flex;
-      align-items: center;
-      cursor: pointer;
-
-      a.btn {
-        width: 100%;
-      }
-    }
-
-    li.submenu {
-      svg {
-        color: #b2b2b2;
-      }
-    }
-
-    li.previous-link {
-
-      padding-left: 11px;
-      span.previous-link {
-        padding-top: 3px;
-      }
-
-      span.back-btn {
-        width: 30px;
-        height: 30px;
-        border-radius: 15px;
-        margin-top: 0px;
-        margin-right: 8px;
-        background-color: #0041DC;
-        color: #fff;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-shrink: 0;
-
-        svg {
-          margin-left: -2px;
-        }
-      }
-    }
-
-    li.router-link-exact-active {
-      a {
-        font-weight: bold;
-        color: #0041DC;
-      }
-    }
-
-    li.previous-link {
-      span.router-link-exact-active {
-        a {
-          font-weight: bold;
-          color: #0041DC;
-        }
-      }
-    }
-
-    li {
-      a.router-link-active {
-        font-weight: 600;
-      }
-    }
-
-    div.subitem {
-      .btn, .btn-link-link {
-        padding-top: 5px;
-        padding-bottom: 5px;
-      }
-    }
-
-  }
-
-  .search {
-    padding: 10px 20px 10px 30px;
-  }
-
-  .muokkaa-kappaleita {
-    padding: 0 20px 0 20px;
-    background-color: #fff;
-
-    // sticky initial state
-    position: fixed;
-    bottom: 0;
-    width: $sidebar-width;
-
-    .inner {
-      border-color: #D8D8D8;
-      border-style: solid;
-      border-width: 1px 0 0 0;
-      padding: 10px;
-    }
-
-    a.router-link-active {
-      color: #888888;
-    }
-  }
-
-  .hallintapaneeli {
-    margin: 0 40px 10px 40px;
-    border-color: #D8D8D8;
-    border-style: solid;
-    border-width: 0 0 1px 0;
-
-    span {
-      padding: 10px;
-      color: $black;
-      display: block;
-    }
-  }
-
-}
